tiny.el (tiny-mapconcat-parse): Less false positives
After this chage, tiny won't try to expand this (point is at |):

    (+ my-variable-1 asdf|)

Previously, it would be interpretted as ("m", "y-variable-", "1").
Now, the restriction is that if you want it to be interpretted like
that, put the cursor right after my-variable-1.

One new restriction is that the separator can't contain digits.  It
didn't make sense before anyway, now it's enforced.